+++ engine/include/rage1/memory.h
@@ -37,9 +37,11 @@ void init_memory(void);
     // all banked functions must be declared as
     //   void function( void );
     typedef void (*banked_function_t)( void );
+    typedef void (*banked_function_arg16_t)( uint16_t arg );
 
-    // trampoline function to call banked functions
+    // trampoline functions to call banked functions
     void memory_call_banked_function( uint8_t function_id );
+    void memory_call_banked_function_arg16( uint8_t function_id, uint16_t arg );
 
     // banked function IDs
     #define BANKED_FUNCTION_SOUND_PLAY_PENDING_FX_ID	0
@@ -47,10 +49,21 @@ void init_memory(void);
     #define BANKED_FUNCTION_ENEMY_ANIMATE_AND_MOVE_ALL	2
     #define BANKED_FUNCTION_BULLET_ANIMATE_AND_MOVE_ALL	3
     #define BANKED_FUNCTION_BULLET_ADD			4
+    #define BANKED_FUNCTION_INIT_TRACKER		5
+    #define BANKED_FUNCTION_TRACKER_SELECT_SONG		6
+    #define BANKED_FUNCTION_TRACKER_START		7
+    #define BANKED_FUNCTION_TRACKER_STOP		8
+    #define BANKED_FUNCTION_TRACKER_DO_PERIODIC_TASKS	9
+    #define BANKED_FUNCTION_INIT_TRACKER_SOUND_EFFECTS	10
+    #define BANKED_FUNCTION_TRACKER_PLAY_FX		11
+    #define BANKED_FUNCTION_TRACKER_REWIND		12
+
+    // WARNING! Don't forget to update the banked function call table
+    // in engine/banked_code/128/00main.asm !!!!
 
     // maximum assigned banked function ID. Keep in sync with the previous IDs
     // 128K versions
-    #define BANKED_FUNCTION_MAX_ID		4
+    #define BANKED_FUNCTION_MAX_ID		12
 
     // Banked function call macros (128K versions) - In 128K mode we
     // redefine the regular calls to banked functions as calls to
@@ -60,6 +73,14 @@ void init_memory(void);
     #define enemy_animate_and_move_all()	( memory_call_banked_function( BANKED_FUNCTION_ENEMY_ANIMATE_AND_MOVE_ALL ) )
     #define bullet_animate_and_move_all()	( memory_call_banked_function( BANKED_FUNCTION_BULLET_ANIMATE_AND_MOVE_ALL ) )
     #define bullet_add()			( memory_call_banked_function( BANKED_FUNCTION_BULLET_ADD ) )
+    #define init_tracker()			( memory_call_banked_function( BANKED_FUNCTION_INIT_TRACKER ) )
+    #define tracker_select_song(a)		( memory_call_banked_function_arg16( BANKED_FUNCTION_TRACKER_SELECT_SONG, (a) ) )
+    #define tracker_start()			( memory_call_banked_function( BANKED_FUNCTION_TRACKER_START ) )
+    #define tracker_stop()			( memory_call_banked_function( BANKED_FUNCTION_TRACKER_STOP ) )
+    #define tracker_do_periodic_tasks()		( memory_call_banked_function( BANKED_FUNCTION_TRACKER_DO_PERIODIC_TASKS ) )
+    #define init_tracker_sound_effects()	( memory_call_banked_function( BANKED_FUNCTION_INIT_TRACKER_SOUND_EFFECTS ) )
+    #define tracker_play_fx(a)			( memory_call_banked_function_arg16( BANKED_FUNCTION_TRACKER_PLAY_FX, (a) ) )
+    #define tracker_rewind()			( memory_call_banked_function( BANKED_FUNCTION_TRACKER_REWIND ) )
 
 #endif